+++ instrumentation/opentelemetry-instrumentation-botocore/src/opentelemetry/instrumentation/botocore/extensions/bedrock_utils.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 
 import json
 from os import environ
-from typing import Callable, Dict, Union
+from typing import Any, Callable, Dict, Union
 
 from botocore.eventstream import EventStream, EventStreamError
 from wrapt import ObjectProxy
@@ -49,8 +49,11 @@ class ConverseStreamWrapper(ObjectProxy):
         self._stream_done_callback = stream_done_callback
         self._stream_error_callback = stream_error_callback
         # accumulating things in the same shape of non-streaming version
-        # {"usage": {"inputTokens": 0, "outputTokens": 0}, "stopReason": "finish"}
+        # {"usage": {"inputTokens": 0, "outputTokens": 0}, "stopReason": "finish", "output": {"message": {"role": "", "content": [{"text": ""}]}
         self._response = {}
+        self._message = None
+        self._content_buf = ""
+        self._record_message = False
 
     def __iter__(self):
         try:
@@ -64,10 +67,17 @@ class ConverseStreamWrapper(ObjectProxy):
     def _process_event(self, event):
         if "messageStart" in event:
             # {'messageStart': {'role': 'assistant'}}
+            if event["messageStart"].get("role") == "assistant":
+                self._record_message = True
+                self._message = {"role": "assistant", "content": []}
             return
 
         if "contentBlockDelta" in event:
             # {'contentBlockDelta': {'delta': {'text': "Hello"}, 'contentBlockIndex': 0}}
+            if self._record_message:
+                self._content_buf += (
+                    event["contentBlockDelta"].get("delta", {}).get("text", "")
+                )
             return
 
         if "contentBlockStop" in event:
@@ -78,6 +88,14 @@ class ConverseStreamWrapper(ObjectProxy):
             # {'messageStop': {'stopReason': 'end_turn'}}
             if stop_reason := event["messageStop"].get("stopReason"):
                 self._response["stopReason"] = stop_reason
+
+            if self._record_message:
+                self._message["content"].append({"text": self._content_buf})
+                self._content_buf = ""
+                self._response["output"] = {"message": self._message}
+                self._record_message = False
+                self._message = None
+
             return
 
         if "metadata" in event:
@@ -91,6 +109,7 @@ class ConverseStreamWrapper(ObjectProxy):
                     self._response["usage"]["outputTokens"] = output_tokens
 
             self._stream_done_callback(self._response)
+
             return
 
 
@@ -112,8 +131,11 @@ class InvokeModelWithResponseStreamWrapper(ObjectProxy):
         self._model_id = model_id
 
         # accumulating things in the same shape of the Converse API
-        # {"usage": {"inputTokens": 0, "outputTokens": 0}, "stopReason": "finish"}
+        # {"usage": {"inputTokens": 0, "outputTokens": 0}, "stopReason": "finish", "output": {"message": {"role": "", "content": [{"text": ""}]}
         self._response = {}
+        self._message = None
+        self._content_buf = ""
+        self._record_message = False
 
     def __iter__(self):
         try:
@@ -159,15 +181,27 @@ class InvokeModelWithResponseStreamWrapper(ObjectProxy):
             #     "inputTokenCount":9,"outputTokenCount":128,"invocationLatency":3569,"firstByteLatency":2180
             # }
             self._process_invocation_metrics(invocation_metrics)
+
+            # transform the shape of the message to match other models
+            self._response["output"] = {
+                "message": {"content": [{"text": chunk["outputText"]}]}
+            }
             self._stream_done_callback(self._response)
 
     def _process_amazon_nova_chunk(self, chunk):
         if "messageStart" in chunk:
             # {'messageStart': {'role': 'assistant'}}
+            if chunk["messageStart"].get("role") == "assistant":
+                self._record_message = True
+                self._message = {"role": "assistant", "content": []}
             return
 
         if "contentBlockDelta" in chunk:
             # {'contentBlockDelta': {'delta': {'text': "Hello"}, 'contentBlockIndex': 0}}
+            if self._record_message:
+                self._content_buf += (
+                    chunk["contentBlockDelta"].get("delta", {}).get("text", "")
+                )
             return
 
         if "contentBlockStop" in chunk:
@@ -178,6 +212,13 @@ class InvokeModelWithResponseStreamWrapper(ObjectProxy):
             # {'messageStop': {'stopReason': 'end_turn'}}
             if stop_reason := chunk["messageStop"].get("stopReason"):
                 self._response["stopReason"] = stop_reason
+
+            if self._record_message:
+                self._message["content"].append({"text": self._content_buf})
+                self._content_buf = ""
+                self._response["output"] = {"message": self._message}
+                self._record_message = False
+                self._message = None
             return
 
         if "metadata" in chunk:
@@ -200,6 +241,13 @@ class InvokeModelWithResponseStreamWrapper(ObjectProxy):
 
         if message_type == "message_start":
             # {'type': 'message_start', 'message': {'id': 'id', 'type': 'message', 'role': 'assistant', 'model': 'claude-2.0', 'content': [], 'stop_reason': None, 'stop_sequence': None, 'usage': {'input_tokens': 18, 'output_tokens': 1}}}
+            if chunk.get("message", {}).get("role") == "assistant":
+                self._record_message = True
+                message = chunk["message"]
+                self._message = {
+                    "role": message["role"],
+                    "content": message.get("content", []),
+                }
             return
 
         if message_type == "content_block_start":
@@ -208,10 +256,14 @@ class InvokeModelWithResponseStreamWrapper(ObjectProxy):
 
         if message_type == "content_block_delta":
             # {'type': 'content_block_delta', 'index': 0, 'delta': {'type': 'text_delta', 'text': 'Here'}}
+            if self._record_message:
+                self._content_buf += chunk.get("delta", {}).get("text", "")
             return
 
         if message_type == "content_block_stop":
             # {'type': 'content_block_stop', 'index': 0}
+            self._message["content"].append({"text": self._content_buf})
+            self._content_buf = ""
             return
 
         if message_type == "message_delta":
@@ -228,6 +280,12 @@ class InvokeModelWithResponseStreamWrapper(ObjectProxy):
                 "amazon-bedrock-invocationMetrics"
             ):
                 self._process_invocation_metrics(invocation_metrics)
+
+            if self._record_message:
+                self._response["output"] = {"message": self._message}
+                self._record_message = False
+                self._message = None
+
             self._stream_done_callback(self._response)
             return
 
@@ -239,7 +297,7 @@ def genai_capture_message_content() -> bool:
     return capture_content.lower() == "true"
 
 
-def message_to_event(message, capture_content):
+def message_to_event(message: dict[str, Any], capture_content: bool) -> Event:
     attributes = {GEN_AI_SYSTEM: GenAiSystemValues.AWS_BEDROCK.value}
     role = message.get("role")
     content = message.get("content")
@@ -253,3 +311,67 @@ def message_to_event(message, capture_content):
         attributes=attributes,
         body=body if body else None,
     )
+
+
+class _Choice:
+    def __init__(
+        self, message: dict[str, Any], finish_reason: str, index: int
+    ):
+        self.message = message
+        self.finish_reason = finish_reason
+        self.index = index
+
+    @classmethod
+    def from_converse(
+        cls, response: dict[str, Any], capture_content: bool
+    ) -> _Choice:
+        orig_message = response["output"]["message"]
+        if role := orig_message.get("role"):
+            message = {"role": role}
+        else:
+            # amazon.titan does not serialize the role
+            message = {}
+        if capture_content:
+            message["content"] = orig_message["content"]
+        return cls(message, response["stopReason"], index=0)
+
+    @classmethod
+    def from_invoke_amazon_titan(
+        cls, response: dict[str, Any], capture_content: bool
+    ) -> _Choice:
+        result = response["results"][0]
+        if capture_content:
+            message = {"content": result["outputText"]}
+        else:
+            message = {}
+        return cls(message, result["completionReason"], index=0)
+
+    @classmethod
+    def from_invoke_anthropic_claude(
+        cls, response: dict[str, Any], capture_content: bool
+    ) -> _Choice:
+        if capture_content:
+            message = {
+                "content": response["content"],
+                "role": response["role"],
+            }
+        else:
+            message = {"role": response["role"]}
+
+        return cls(message, response["stop_reason"], index=0)
+
+    def _to_body_dict(self) -> dict[str, Any]:
+        return {
+            "finish_reason": self.finish_reason,
+            "index": self.index,
+            "message": self.message,
+        }
+
+    def to_choice_event(self, **event_kwargs) -> Event:
+        attributes = {GEN_AI_SYSTEM: GenAiSystemValues.AWS_BEDROCK.value}
+        return Event(
+            name="gen_ai.choice",
+            attributes=attributes,
+            body=self._to_body_dict(),
+            **event_kwargs,
+        )